Add mock AlgorithmImports and update project metadata
- Introduced `AlgorithmImports.py` with mock classes for local development, including `QCAlgorithm`, `MockPortfolio`, `MockSymbol`, and `MockOption`.
- Updated `pyproject.toml` to ensure proper formatting and added missing newline.
- Created `uv.lock` for dependency management.
- Added `PKG-INFO`, `entry_points.txt`, `requires.txt`, and `SOURCES.txt` for package metadata and dependencies.
- Enhanced `risk_manager.py` to ensure consistent float return types for volatility and PnL calculations.
- Added new files for dependency links and entry points to facilitate package installation and execution.
- Updated `main.py` and `position_manager.py` to integrate new imports and ensure compatibility with the mock framework.

+++ core/risk_manager.py
@@ -201,7 +201,7 @@ class RiskManager:
             
         # Calculate volatility from recent PnL data
         recent_pnl: List[float] = all_daily_pnl[-self.volatility_lookback:]
-        volatility: float = np.std(recent_pnl)
+        volatility: float = float(np.std(recent_pnl))
         
         # Adjust position size based on volatility regime
         if volatility > self.volatility_threshold:
@@ -251,7 +251,7 @@ class RiskManager:
         completed_trades = [t for t in all_trades if 'pnl' in t and t['pnl'] > 0]
         if not completed_trades:
             return 100  # Default assumption
-        return np.mean([t['pnl'] for t in completed_trades])
+        return float(np.mean([t['pnl'] for t in completed_trades]))
     
     def get_average_loss(self) -> float:
         """
@@ -270,7 +270,7 @@ class RiskManager:
         completed_trades = [t for t in all_trades if 'pnl' in t and t['pnl'] < 0]
         if not completed_trades:
             return 200  # Default assumption
-        return abs(np.mean([t['pnl'] for t in completed_trades]))
+        return abs(float(np.mean([t['pnl'] for t in completed_trades])))
     
     def should_stop_trading(self):
         """
